[IMP] stock@7.saas~5: finally solve all stock's issues
Note: the quantities after migration may vary in product.product or
product.template but this is normal. We increased the precision of the
UoM of the product, so the calculation gives a better result compared to
Odoo 7.0.

According to PBO, we should deliver with this difference since it's
better than before.