[FIX] stock@7.saas~5: rounding on minimal float precision and...
It seems that Postgres fails on this test:
[local] template1=# SELECT (28.000 / 12.000 * 12.000) = 28;
 ?column?
----------
 f
(1 row)

But if you round to the minimal precision representable by the type
"double precision", it works fine:

[local] template1=# SELECT round((28.000 / 12.000 * 12.000), 15) = 28;
 ?column?
----------
 t
(1 row)

http://www.postgresql.org/docs/9.0/static/datatype-numeric.html

Other fix: prevent testing the conversion of stock moves to the default
UoM (product.template's UoM) if they are the same. This is useless.