[IMP] speed up procurement groups creation if parameter is present
if the 'migration.stock.create-proc-only-running' parameter is present,
we only create procurement groups from sale orders not in 'done',
'cancel', 'draft' state, otherwise we create them from sale orders not
in 'draft' state.

This parameter is useful for migrations taking several days. In this
case, set the 'migration.stock.create-proc-only-running' parameter to
True.