[IMP] account@8.saas~6.1.1: bootstrap account.invoice.line price_subtotal_signed
Recomputing it using the ORM system forces writing
on every account.invoice.line in the system, usually
taking hours if there are a couple of thousand lines,
due to the many cascading dependencies and
recomputations.
Bootstrapping it avoids this huge cost + the risk of
having side-effects on other columns due to various
recomputations. (e.g. altering past invoices due to
behavior differences in the new code)