[FIX] {base,account}@9.0: force install of `account_tax_python` in base
The module loading is done in (at least) two passes: first only the base
module, then the other ones (in a loop until there is new modules to load).
By forcing the installation of `account_tax_python` inside the
`account` module, it is loaded in a third pass, after other modules
that may need it loaded, like `sale`.