[IMP] mrp@saas~12: add first phase of mrp migration, won't support open production orders for the moment
- In pre, we need to remove the necessary views, drop workflows on work orders and manufacturing orders
- In post, we have a file to generate the manufacturing picking types and assign them to mos and procurement rules.
	We also need to change the state of the open production orders (not needed yet)
- In a second post file, we update the link between consumed and produced quants which was done with the consumed_for field on the move before (traceability)
There is a detail of the move object added (stock.move.lots) which need to be created every time we used a lot and the quantity consumed/produced
need to be set on the move too.
The routing on the production order became a related with the bom, so if non-existing combination was chosen, it should duplicate the bom.
Workorders from done/cancelled production orders, should be set to done/cancelled too.