[FIX] project_issue: common fields with project issue
When doing the jump from 9.0 to saas-10,
therefore skipping the test saas-8/saas-9,
the `store` field of the claim field are not
correctly computed. You therefore cannot
rely on this field to know if the field is stored
or not.

Besides, it wasn't really needed, as just before
there is a join on information_schema, checking that
the field are in database. If these fields are in databases
well, it means they are stored :D.

opw-690004