[IMP] stock_account: migration to v11
In order to have the same stock value before and after migrating
real valuation to fifo, we check how many quants on an incoming are
still in stock and calculated the remaining_qty and remaining_value
with this.  And the value is calculated from the quants moved.

For the landed costs, we need to subtract the amount back from the
value as the landed cost value field will represent that cost.

We should also make sure that only one move gets the valuation of one quant
e.g. if a quant is returned from customers, it went two times through an in.
It is important to put its remaining_qty on the last one as is done here and
not on both. (as this would give a wrong valuation)

Also the internal transit locations are important for valuation