[FIX] stock @11.0: remove constraint instead of playing with not exists
In the extreme case, it is possible that a quant is moved
by two different moves (negative stock, returns, ... you
have to do an effort) in the same picking.  In order to avoid
tracebacks in the migration, we allow to have multiple times
the same entry in the stock_quant_move_rel many2many table by
dropping the unique constraint.