[FIX] stock_account: for negative quants, use propagated_from_id on quant
In v10, there was no accounting entry created
for the stock that became negative.  (only solved when

In v11, there is an accounting entry, which is corrected by the fifo_vacuum algo.

Another problem we had before this fix is that a negative
quant would not be taken into account when it is propagated one
is out, but it was itself not created by the out.
(by a step before with not enough stock)

So, the solution is to search for work with the positive quants
that were propagated from a negative quant (propagated_from_id not null)
and to count them in for the remaining_qty but leave the remaining_value
0 as they were not accounted for in v10.  When launching the valuation
report in v11 it will apply financial fifo, so the negative remaining_qty
might immediately be reconciled with the fifo_vacuum.

We need to take the last out for the case the quant returned in stock and was sent
back to the customer.