[FIX] base@9.0: prevent NULL value for type field on partners
The type field is used in the 'private address' security rule that
prevents people not in the required group to see private addresses.

The domain for that rule is [('type', '!=', 'private')], which does
not take into account NULL values at the SQL level because expression.py
does not include OR IS NULL conditions in string comparison with a !=
operator.

Although the field is not required and has no default, the private
addess feature was introduced in 9.0 so I would like migrations to that
version to support it... and partner of undefined type are a bit weird
functionnally speaking, so I don't thing forcing this change here is
bad.