[IMP] stock7.saas~5: allow override of minimal rounding value
This assertion may make sense when the database is coherent/restrained
in its usage of uoms and factors, but in some case where a user has
encoded factors such as 1/33 and sold 1.538576646362858 unit of those,
this precision will never be reached.

Let's reduce the default precision by 2 factors and allow an environment
variable override when it's the case.

opw-1840326