[IMP] purchase@saas~11.5: some things can't be computed
If the uoms of the purchase line and product template are not in the same category,
then the recompute will fail. Given that having a mess of uom categs is pretty much
a given for any database older than, I don't know, 2 months, let's be a bit preventive
here.

If the uoms are not compatible, we behave in the same way as `_compute_quantity`
would if `raise_if_failure` were set to `False`: we just use the same quantity.
This new field seems to be there for reporting only, and fixing all past entries because
of uoms problems is not realistic anyway.