[FIX] account@saas~11.4: clean up 9.0 fake record
A 'fake' CoA was introduced in 8.0->9.0 migrations so that customers
could not apply a second chart of account to a company that already had
one set up - but without a clear link to what was actually applied in
8.0, we needed a fake record.

Fine by me, but in 11.2 and 11.4 we introduced new required fields on
that model and this record need some new values to allow postgres to
add the NOT NULL constraint properly.

Add some meaningless default values, the customers will never actually
see them since it's impossible to apply this fake CoA to new companies.