[FIX] base@10.0: ensure migration of pos_blackbox_be module can continue
The 'pos_blackbox_be' module is only certified in v11.0
We need to ensure that the migration process continue when the target
version is 11.0 but continue blocking it if the target version is 10.0

At the end of the upgrade to v10.0, that module will be in 'to upgrade'
state if it was initialy installed because that module is "installable
= False" in v10
If the target version is 10.0, this is OK. It will block the migration
process. That's intended.
But if the target version is 11.0, we set it as 'installed' so that the
migration can continue to v11.0.