[FIX] saas-12.3: account: taxes migration: fix multi-l10n issue
Deleting the backup financial report lines for which there was no mapping with the new model in case there existed such a mapping was wrong. Indeed, if a database had two companies with different l10n, and one of these l10n had such a mapping and the other one used financial tags (be and es for example), financial tags were not assigned to the migrated taxes, since the corresponding financial report lines had been removed from backup table.

[IMP] saas-12.3: account: taxes migration: Make sure 0% taxes and fixed taxes with amount=0 have an empty taxe repartition

This is done to better match the templates defined in l10n modules, and avoid noise in the warnings shown when a template does not correspond to a migrated tax instance.

[FIX] saas-12.3: l10n_ca: fix xmlid matching in sql query, avoiding matching something else by mistake