[FIX] account@saas~12.3: avoid 0% tax lines when merging groups of taxes
It is possible to encounter a 0% tax line, because of Miscellaneous Operations, the way POS computes taxes or simply old data gotten from previous versions. It crashed the script when trying to merge the corresponding child taxes into their parents. Now, we only remove the tax_line_id from these lines before merging them (not a problem since, they don't have an impact on the accounting, as their balance is 0).

Closes #621 
Fixes #613