[FIX] hr_holidays@saas~11.2: type_value constraint prevents to add new value in selection holiday_type
Request 40424
A custom module add the value `supervisor` in the possible values
of the selection field `hr.leave.holiday_type`.

Prevent the migration to stop because the raise of this constraint by
removing the constraint and attempt to re-add it afterwards.
In the case of this upgrade request,
the attempt to reset it will fail, but without blocking.
This is actual the default behavior of Odoo regarding the addition
of constraints. In this migration script it's not this behavior
because the model table is created first, with the constraints already set,
and then filled with the data of another table,
which will violate the constraint.

Now we fill the table with the migrated data,
and then we attempt to add the constraint.

Then it's the responsibility of the custom module
to alter the constraint so it deals with the possible
value added in the selection field, and the constraint
will be reset automatically with the ORM with a simple
-u module_name