[FIX] base@10.saas~17: Don't stop upgrade because of deprecated reports
Before this revision, having deprecated reports (rml2pdf, webkit, ...)
blocked the upgrade process.

From this revision, it simply adds a message in the migration report
send to admins at the end of the upgrade process to warn them
they had reports that will no longer work. In addition,
when removing the related fields, we avoid to actually remove
the table columns, to keep the data in case the user would like
to recover it somehow.