[FIX] base@0.0.0: better logs when disabling/deleting custom view
When a custom view fails its validation,
according if it comes from a module
or if the view was created in database directly,
we either just disable it (`active = False`) or
we delete it
(as it should be modified in the XML file and then re-imported with a `-u`)

The log was just stating "disabling" even when the view was completely
deleted. It now correctly state "deleted" when the view is deleted,
and what was its xmlid and in which file it was.

I also take the opportunity to improve the migration report regarding the altered views:
 - Display the xml_id of the deleted views, as well as their arch_fs
   so the customer or our support employees can easily know from which file was defined the view,
 - No need for a link to the view in the case of a deleted view, obviously.