[FIX] survey@saas~12.2: fix survey resequencing
Sequences in survey pages and questions may be left all to zeros, in
which case they were ordered by the id (_order = 'sequence,id').

The issue with the resequence SQL script was that in such a situation,
it was grouping the pages to the top, before all the questions.

This fix fallbacks on the id if the sequences make no difference.