[FIX] base@saas~12.5: correct db-specific script
Previous script was putting an inexistent class `o_half_screen_height`
on event_event.cover_properties.

What we want is actually keep the `cover_mid`, which tells the size of
the banner if there is a background picture, and just add the
`o_record_has_cover` class if there is a background picture.

The new request is also using postgresql's json operators to make the
changes.

closes odoo/upgrade#741

Signed-off-by: Christophe Simonis <chs@odoo.com>