[FIX] base@0.0.0: force update of group_system
This is related to
odoo/odoo@81d1296323fdaeb88c37bde45f7443afd6ebd0ac

When a user has the Administration / Settings group,
but is not an accounting advisor,
attempting to access the Accounting settings raises
an access error.

This shouldn't be possible to be a settings admin
without being an accounting advisor thanks to the above
commit, which makes Administration / Setings imply
Accounting / Advisor,
except
 - if you removed this implies from Administration / Settings
   and then set noupdate to True on this group
 - if you come from a db < 11.0 (e.g. 10.0) and set noupdate to
   True on this group. Then you never had the chance to get the implieds
   during upgrades.

This is the case of the database subject of the opw ticket 2183312:
Administration / Settings is set to noupdate
and doesn't have the imply on Accounting / Advisor.
When the admin (ID 2) tries to access the website configuration,
it loads the accounting configuration,
and raises the given access error.

opw-2183312

closes odoo/upgrade#749

Signed-off-by: Christophe Simonis <chs@odoo.com>