[FIX] base@saas~12.5
Payment term use in multiple company and for the online payment belong to odoo SA but are user by other company. This cause in v13.0 to break the check on company consistency during online payment flow amongs other things

It's safer to remove the company_id on all payment term in the migration script and then set back manually some company_id on company specific payment term than the opposite.

closes odoo/upgrade#755

Signed-off-by: Nicolas Seinlet (nse) <nse@odoo.com>