[FIX] base@0.0.0: check and fix custom view one by one
Multiple views are passed in the `self` of `_check_xml`,
and they can be completely unrelated,
e.g. a form view on res.partner, a kanban on project.project, and so on.

The algorithm was disabling the view one by one until all validation
passed, but it could happen it was disabling a custom kanban view
from project.project while the validation failed on a
form view on res.partner.

To overcome this issue, validates the views one by one,
so when one fails, it tries to disable custom views
related to this one, and not completely unrelated
views.

See https://github.com/odoo/odoo/blob/11.0/odoo/addons/base/ir/ir_ui_view.py#L1331-L1346

closes odoo/upgrade#756

Signed-off-by: Denis Ledoux (dle) <dle@odoo.com>