[IMP] website_forum@saas~12.5: give a name to all posts
From Odoo 13.0, the list of answers of a user displays the name of the
answer itself, and not its parent's.
Answers created directly on a question are given a name, but answers
created from a comment are not.

This commit gives a name to the forum posts that did not have one
because they were created from a comment.

closes odoo/upgrade#769

Signed-off-by: Christophe Simonis <chs@odoo.com>