[IMP] base@saas~13.4: add dep 'partner_autocomplete' to 'crm_iap_lead(_enrich)'
To ease the re-usage and lessen code duplication, we define a base iap
information template in partner_autocomplete that we use everywhere it's
needed.

'partner_autocomplete' was most likely already installed since it's auto on:
['web', 'mail', 'iap'].
We add it as explicit requirement since we now use one of its templates.

LINKS

Task ID 2081471
Community PR odoo/odoo#40349

closes odoo/upgrade#775

Signed-off-by: Nicolas Seinlet (nse) <nse@odoo.com>