[IMP] 12-13: account: avoid warning with l10n_generic_coa when migrating standard taxes
When migrating the two 15% taxes from l10n_generic_coa, a warning was logged telling the migrated repartition for theses taxes did not match the template. This warning was always reported, despite the migration being correct, because of a change in the account templates in v13 (saas-12.5, here https://github.com/odoo/odoo/commit/4ddce229c49a02e27664e4228fc3c992645cb5e3 ). Though it was not a problem on customer dbs, it was annoying on runbot, for migration tests.

closes odoo/upgrade#783

Signed-off-by: Christophe Simonis <chs@odoo.com>