[REF] crm@saas~13.3, website_crm_score@saas~13.3: clean models and update statistics fields
Purpose of this commit is to clean statistics fields definition on crm.team.
Future tasks will improve the use of sales team in crm and first steps is to
clean a bit models linked to sales team.

Purpose of this commit is to clean models in website crm score. Future tasks
will improve the use of sales team in crm and first steps is to clean a bit
models linked to sales team.

Purpose is to achieve a better code understanding with better field names.
Some renaming and code linting is also performed according to guidelines.

LINKS

Task ID 2088565
Community PR odoo/odoo#41955
Enterprise PR odoo/enterprise#7322

closes odoo/upgrade#781

Signed-off-by: Nicolas Seinlet (nse) <nse@odoo.com>