[FIX] sale@saas~11.5: rename xmlid `website_quote.so_quotation`
The xmlid
`website_quote.so_quotation`
has been renamed to
`sale.sale_order_portal_template`
at odoo/odoo@4c31a758cf0715b1c1ff2f5b3cef18fec92c2d84

https://github.com/odoo/odoo/commit/4c31a758cf0715b1c1ff2f5b3cef18fec92c2d84#diff-b0150c7e2ec8a594fc21ecfb361f7369R141
https://github.com/odoo/odoo/commit/4c31a758cf0715b1c1ff2f5b3cef18fec92c2d84#diff-679ce32bacca6ca00054ccd05caab2fdL248

The arch is a bit different, but some blocks are similar. You can for instance find the below in both views in the diff:
```xml
<strong>This offer expired!</strong><br/>
Contact us for new quotation.
```

Without this, the view is deleted at the end of the upgrade,
during the deletion of the ir.model.data,
(as its not deleted by a upgrade script with util.remove_view)
and if the database has custom views,
it fails because there are still inheriting views referencing this one.

In the case of the database with which has been detected this issue,
compactor-management-company.odoo.com,
the inheriting view even works.

This commit has been introduced in Odoo saas-11.5.

There is another upgrade script in saas-11.5 which renames
the module `website_quote` to `sale_quotation_builder`
https://github.com/odoo/upgrade/blob/master/migrations/base/saas~11.5.1.3/pre-10-modules.py#L34
which is launched before this one, because its on the module `base`,
hence the fact we rename from `sale_quotation_builder` instead of `website_quote`.

closes odoo/upgrade#890

Signed-off-by: Christophe Simonis <chs@odoo.com>