[FIX] sale_management@saas~11.5: rename xmlid `website_quote.view_sale_quote_template_form`
The xmlid
`website_quote.view_sale_quote_template_form`
has been renamed to
`sale_management.sale_order_template_view_form`
at odoo/odoo@4c31a758cf0715b1c1ff2f5b3cef18fec92c2d84

https://github.com/odoo/odoo/commit/4c31a758cf0715b1c1ff2f5b3cef18fec92c2d84#diff-7ef6d013bc9245e83fb4b8dbd8ca3cf4L12
https://github.com/odoo/odoo/commit/4c31a758cf0715b1c1ff2f5b3cef18fec92c2d84#diff-698d74a85aa4e1ff32fbe79795fdff75R13

The arch is a bit different, but some blocks are similar. You can for instance find the below in both views in the diff:
```xml
<div id="number_of_days">
    <field name="number_of_days" class="oe_inline"/> days
</div>
```

Without this, the view is deleted at the end of the upgrade,
during the deletion of the ir.model.data,
(as its not deleted by an upgrade script with util.remove_view)
and if the database has custom views,
it fails because there are still inheriting views referencing this one.

In the case of the database with which has been detected this issue,
nobus.odoo.com,
the inheriting view even works.

This commit has been introduced in Odoo saas-11.5.

There is another upgrade script in saas-11.5 which renames
the module `website_quote` to `sale_quotation_builder`
https://github.com/odoo/upgrade/blob/master/migrations/base/saas~11.5.1.3/pre-10-modules.py#L34
which is launched before this one, because its on the module `base`,
hence the fact we rename from `sale_quotation_builder` instead of `website_quote`.

closes odoo/upgrade#892

Signed-off-by: Christophe Simonis <chs@odoo.com>