[FIX] website@saas~13.3: remove wrong field definition table for res_partner.visitor_ids.
visitor.partner is a many2one to partner. But partner.visitor_ids is a
many2many. The result is when you set the partner_id on visitor, the (what
should be) inverse field is not set accordingly.

This inverse field should be a one2many fields.
This commit fixes the field definition.

Note : This fix is done in master even if the field definition is wrong in v13.
But as this relation is not used anywhere (yet), this is not a 'must be' fix
for v13.

See odoo/odoo#40577

closes odoo/upgrade#905

Signed-off-by: Christophe Simonis <chs@odoo.com>