[REM] data_merge@saas~13.3: convert selection field from static to dynamic
Field was converted from a static selection (`Selection(list)`) to a
dynamic one (`Selection(callable)`).

The latter doesn't have ir.model.fields.selection records associated
with its values so the ORM attempts to remove those records... and set
to null the fields using those values in the host model, which is
undesirable and inconvenient.

Delete the selection records bypassing the ORM so we can keep the
values in the database.

closes odoo/upgrade#930

Related: odoo/enterprise#9159
Signed-off-by: Nicolas Seinlet (nse) <nse@odoo.com>