[FIX] product@saas~12.2: inject cron to resize images post upgrade
In most cases,
the filestore is not provided when upgrading databases.

Therefore, it's not possible to resize images stored in filestore
during the upgrade.

To overcome the issue, we inject a cron which will resize the images
automatically once the database is put back in production.

The cron will run every hour, resizing 1000 images at a time.

It's nevertheless not possible for the cron to auto-destroy itself
once its done, or to disable itself, because of the `lock_cr`
cursor which take a lock on the cron record, therefore
preventing the cron to destroy itself or write on itself.

Instead, we will automatically remove the cron during the next upgrade.

e.g.
A database is upgraded from 12.0 to 13.0, we inject the cron.
A year later, the database upgrades from 13.0 to 14.0, we remove the cron.

closes odoo/upgrade#727

Signed-off-by: Christophe Simonis <chs@odoo.com>
Co-authored-by: Olivier Laurent <olt@odoo.com>