[IMP] base@saas~13.3: remove "view_partner_short_form" view
This commit explicitly removes the view_partner_short_form view on partner
defined in base. Indeed this view is not referenced in any action or code
and has an higher priority than the default form view, meaning it is not
used anymore.

As it has several inheriting views and could have been customized due to
previous use of that view a pre-migrate hook is added to remove it explicitly.

Related PR has other view removed but as those are not important Odoo default
behavior for their removal is considered as sufficient.

Task ID 2196317
PR odoo/odoo#48086

closes odoo/upgrade#966

Signed-off-by: Nicolas Seinlet (nse) <nse@odoo.com>