[FIX] social_linkedin_company_support@13.0: add the module
Purpose
=======
Before, users could publish on their personal LinkedIn account, but not
on their company pages.

Now that we have the permissions, we prefer the users to use their
company pages, so we can prepare the next iteration of the `social_linkedin`
module that will allow you to read the post of your company pages.

We do not want existing users to keep their personal account, so we add a
technical module to minimize impact and support.

Task-2156000
See odoo/enterprise/pull/9388

closes odoo/upgrade#985

Signed-off-by: Nicolas Seinlet (nse) <nse@odoo.com>