[FIX] base@0.0.0: match_uniq for non-stored fields from 13.0
`_write`, which is used to bypass the protection preventing
to write on `code`of `res.lang`,
does not work for non-stored fields in 13.0.

`image` on `res.country` is considered as a non-stored field
because its a binary stored in the filestore.
Therefore, before this revision,
the `match_uniq` on `res.country` failed from 13.0.

e.g
- Create a database in 12.0, installing just base,
- then
```
DELETE FROM ir_model_data where model = 'res.country' and name = 'fo';
```
- then upgrade to 13.0.

closes odoo/upgrade#988

Signed-off-by: Christophe Simonis <chs@odoo.com>