[FIX] base@0.0.0: _check_xml with multiple custom views on the same model
`_validate_custom_views` pass multiple views in self.
However, this override was ending (with `return`)
as soon as the first view in `self` was validated.

Obviously, the other views in `self` should be validated,
not just the first one.

The issue wasn't that severe, because `_validate_custom_views`
calls `_check_xml` model by model,
and most of the time there wasn't many custom views
on the same model,
and when there was, they were validated most of the time
because the validator goes back to the root, then build the views
using all children, then validates.

But views completely redefining view for models,
without inheriting from another view,
were not being validated anymore.

closes odoo/upgrade#1027

Signed-off-by: Denis Ledoux (dle) <dle@odoo.com>