[IMP] util.py: add move_field_references helper
The goal of this new helper is to move references linked to a field
when it has been copied to another new field in another model and/or
copied has a new field in a same model.

In this version, this helper just handles ir_server_object_lines and
ir_act_server.

Co-Authored-By: Christophe Simonis <KangOl@users.noreply.github.com>