[ADD] website_sale@saas~13.4: migrate ribbons from product_style
Previously, sale ribbons on products were implemented using a html class
defined on the product_style model. in odoo/odoo#47574 this model was
removed, and a new model for custom ribbons was added. This commit adds
a script to migrate the old sale ribbons to the new sale ribbon.

task-1981170

closes odoo/upgrade#1079

Related: odoo/design-themes#219
Signed-off-by: Nicolas Seinlet (nse) <nse@odoo.com>