[FIX] account@saas~12.4: no matching found between voucher and move lines
The `migrate_voucher_lines` function uses a big set of conditions to try
to match voucher lines to move lines. For the database of the upgrade
request #43412, some voucher lines were matched to 3 move lines, without
being able to select only one of them among all the conditions set.

To handle this case, a new condition has been added at the end of the
conditions set, to use it as fallback.

Upgrade request: 43412

closes odoo/upgrade#1082

Signed-off-by: Nicolas Seinlet (nse) <nse@odoo.com>