[IMP] {base,mass_mailing,link_tracker}@saas~13.4: add fields for day+1 KPI sending
Purpose
=======
Provide a follow-up/overview of the Mailing 24 hours after its been sent.

Add a new field "label" on the link trackers. This field represents the
label of the anchor or of the button of the link tracker in the HTML content.
We do not initialize this field in SQL because it will be used only for future
mailings (and it has no use for existing links).

The field `kpi_mail_required` is used to know if we should send the KPI mail
one day after the sending date or not.

Task-2227411
See odoo/odoo/pull/49836

closes odoo/upgrade#1094

Related: odoo/odoo#49836
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>