[FIX] stock: test on hand quantity for bom kits
The available quantity computation for bom kits changed
in 13.0. It now computes the maximum quantity of that
kit that can be produced.

Therefore, for these products, the quantity can
pass from 0 to something else (1, 2, ...)
when passing from 12.0 to 13.0.

As a solution, we simply ignore bom kits
if the source is before 13.0 and the target after 13.0, included.

closes odoo/upgrade#1137

Signed-off-by: Denis Ledoux (dle) <dle@odoo.com>