[FIX] stock_account@saas~12.4: avoid aml without stock moves
In 81ca568, we created stock valuation layer for each account move line
representing a stock movement. The issue was the join between
account_move_line and stock_move was too permissive.
We was creating svl for each aml regardless of the existence of
any stock_move.

This commit change the left join by an inner join to ensure only the aml
related to a stock_move are taken into account to create the valuation
layers.

closes odoo/upgrade#1161

Opw: 2237164
Signed-off-by: Nicolas Seinlet (nse) <nse@odoo.com>