[FIX] base: test mock crawl, ignore custom fields when reading action
Some custom fields loaded in the registry might not be read-able
e.g. compute field which are not stored.

Reading all the fields when loading an action is not necessary,
only the base fields are. To make the crawler fail even less,
we just read the fields we need.

Unfortunately, this wasnt that easy to do because
the standard method `get_bindings` reads the action
without passing a fields list, and therefore it was reading
all the fields of the action.

I therefore patched read of `ir.actions.actions` to always
read a specific list of fields whatever you ask,
just the time to load the action.

closes odoo/upgrade#1178

Signed-off-by: Denis Ledoux (dle) <dle@odoo.com>