[ADD] base@saas~12.5: `read_group` `fields` arg must be list
Since
https://github.com/odoo/odoo/commit/9595dd06a1d349b2fb84969374d57c32a5faab74#diff-9af8b31d921dcccb7bcf230271fd3ef4R2147-R2148

Passing a string for the `fields` argument of
`read_group` raises an exception.

Actually, the behavior when passing a `string`
for the `fields` arg was already wrong before 13.0,
but it didn't raise: It was just not taken into account at all.

Therefore, this lib2to3 fixer convert all calls to `read_group`
using a simple `string` for the `fields` arg in
the server action or the compute method of manual fields.

This is nevertheless still accepted to pass a `string`
for the `groupby` arg thanks to:
https://github.com/odoo/odoo/blob/9595dd06a1d349b2fb84969374d57c32a5faab74/odoo/models.py#L2105

If one day we want to correct the `groupby` arg as well,
the pattern can become:
```
power< any*
    trailer<
        '.' 'read_group'
    > trailer<
        '('
        arglist<
            any ',' (
                fields=STRING ',' group_by=STRING
                | fields=STRING ',' any
                | any ',' group_by=STRING
            ) any*
        >
        ')'
    >
>
```
At the moment, I decided not to alter it,
to avoid taking even more risks by attempting to migrate code automatically.

closes odoo/upgrade#1194

Signed-off-by: Christophe Simonis <chs@odoo.com>