[FIX] tests: Disable AmountTotalSignedIntegrity on platform
The platform uses the tags `test_invariant` to decide
which unit test should run,
using
`--test-tags integrity_case.test_prepare`

This is not possible to combine tags with an `and` operator
to have both the tags `upgrade` and `integrity_case`.

I therefore remove the tag `integrity_case` from this
unit test to disable it, as it is currently broken.

closes odoo/upgrade#1197

Signed-off-by: Denis Ledoux (dle) <dle@odoo.com>