[FIX] mail@saas~12.5: `notified_partner_ids` may already exists
The many2many field `notified_partner_ids` existed in 8.0 and was removed in 9.0.
https://github.com/odoo/odoo/blob/8.0/addons/mail/mail_message.py#L138
Its table was removed correctly thanks to
https://github.com/odoo/upgrade/blob/master/migrations/mail/9.0.1.0/pre-10-model-changes.py#L129
but its field in the table `ir.model.fields` was not correctly removed in any upgrade script.
If this field is there, and we attempt to rename `needaction_partner_ids` to `notified_partner_ids`
It raise an unique constraint in the `ir_model_fields` table

Request 47822, 47744

closes odoo/upgrade#1228

Signed-off-by: Christophe Simonis <chs@odoo.com>