[FIX] base: crawler test, wrap the action read in the try..except
Sometimes, the action of the menu no longer exists.
This is not normal, but it should not end the test abruptly,
but rather report the failing menu at the end,
when checking the before/after failing menus

In the observed case, the issue is related to
odoo/odoo#51552

The server action `ir_actions_server_website_dashboard` has been deleted.
In the upgrade process,
the server action has been re-created,
but as it was now on the menu `menu_website_dashboard`
rather than on the menu `menu_dashboard`,
the action of the menu `menu_dashboard`
has not been updated/emptied correctly.

Therefore, it still pointed to a no longer existing action.

closes odoo/upgrade#1234

Signed-off-by: Denis Ledoux (dle) <dle@odoo.com>