[IMP] util.create_column: allow to set a default value
util.create_column now accept a keyword-only argument `default` to fill
this new column with the given value.

It's done using the `ALTER TABLE ... ADD COLUMN ... DEFAULT ...`
statement to take advantage of PostgreSQL 11 which do not rewrite all
rows, but store the evaluated value in the table's metadata.
On older PostgreSQL versions, this has the same cost as an `UPDATE`
afterward.
The SQL default value is then removed from the schema to avoid crazy
Odoo behavior.


Co-authored-by: Christophe Simonis <chs@odoo.com>