[FIX] sale@saas~11.5: convert amount from unknown company
For some databases, when upgrading to 11.5, a crash occurs when
recomputing the amount in a sale order line with no related company_id
(which is needed to convert the currency).

A countermeasure already existed for some of these cases, by setting the
company_id of the sale order line to the one from the related sale
order.

However, the problem still exists if there are sale orders with no
company_id, as it is not a required field in 11.5. To prevent the crash
and be able to recompute the values, default values are set on the
company_id for sale orders which do not have one set.

Task ID: 2251382

closes odoo/upgrade#1254

Signed-off-by: Denis Ledoux (dle) <dle@odoo.com>