[FIX] base@13.0: one group type per user
A constraint added in 12.0 implies all users to be in only one group type
(either employee, either portal, either public)

Since odoo/odoo@459e6dc1d296c6fca11984c88046aa17ad32b026 (Odoo 13.0)
The constraint is checked as soon as you had a group
to a user or a user to a group,
for instance through the data xml files.
In addition, the constraint checks the groups of ALL users
as soon as the constraint is called on mutiple users
https://github.com/odoo/odoo/commit/459e6dc1d296c6fca11984c88046aa17ad32b026#diff-c087b09b8455bbf3fa546a56c6e271bcR435

And in some data files, we sometimes had two users to a group in the same call.

Therefore:
 - in 12.0, it was possible to bypass the constraint because of the reasons explained in revision
   odoo/odoo@459e6dc1d296c6fca11984c88046aa17ad32b026
   therefore leading to databases not respecting the constraint
 - in 13.0, as soon as two users are added to a group in a data xml files,
   the constraint is checked on all users

We therefore need in upgrade scripts to make the constraint respected.

Besides,
 - for 13.0, it needs to be done in a pre script,
   because it must be done before the data xml files are loaded,
 - for 12.0, it must be done in a post script, or at least after
   `base/saas~11.5.1.3/pre-20-user_system.py` so the user "2" is created

The initial idea was to put this in a `0.0.0` script,
but it cant be a post `0.0.0` for 13.0,
and it cant be a pre `0.0.0` for 12.0, because `0.0.0` script are executed
before specific version pre scripts, and it requires the script `base/saas~11.5.1.3/pre-20-user_system.py`.
e.g.
`base/0.0.0/pre-30-only-one-user-type-group.py` would be executed before
`base/saas~11.5.1.3/pre-20-user_system.py`

The idea is therefore to execute this cleaning script as a post in 12.0,
and as a pre in 13.0.

In the future, if its still possible to bypass the constraint during the 13.0 lifetime
and we therefore need as well to clean this when upgrading to 14.0, we can
then convert `base/13.0.1.3/pre-30-one-user-type-group.py`
to `base/0.0.0/pre-30-one-user-type-group.py`
with a `if util.version_gte("13.0"):`

Requests:
45366
45326
46901
45207
43428
46067
45521
45233

closes odoo/upgrade#1255

Signed-off-by: Christophe Simonis <chs@odoo.com>