[FIX] base@0.0.0: check view root
When searching to fix a view (to disable/overwite one of the inherited views)
then we loop on all inherited views, whether they are primary or extension

e.g.
https://github.com/odoo/odoo/blob/2905d05fc3d30562c5e7510217a6e97e26e87a37/addons/account/views/account_move_views.xml#L351-L354
This is a primary view, but which inherits from a view anyway.

When searching for an inherited view to disable/overwrite for this view,
we must loop on all inherited views in the chain,
not only up the first `primary` view.
Otherwise we immediately stop on this view.

But, when calling `_check_xml` after a view got disabled/overwritten,
there we must start over the first `primary` view.
Otherwise in the case of `product.template`/`product.product`
we would call `_check_xml` on the common view on model `product.template`
and never validates views on the model `product.product` inheriting on a
`product.template`view.

Request 48158

closes odoo/upgrade#1267

Signed-off-by: Denis Ledoux (dle) <dle@odoo.com>