[FIX] account@saas~12.3: crash on duplicate invoice creation
Before this commit, some databases crash upon upgrading to 12.3 due to
the creation of an invoice with an already existing key.

The problem is caused by the prefixes/suffixes used for the sequence of
some journals in these databases, which does not account for date
ranges. Since the invoice in the migration is created with the current
date, it may trigger a new sequence, but since the prefix/suffix does
not update, it uses an already existing invoice number.

For example, if we have a journal with a sequence that has prefix
'INV/' for the year 2019, invoices for this journal would be named
INV/001, INV/002, etc.
If we then use the same journal to create an invoice for 2020, the first
invoice would be named INV/001 again, since there is no date formatting
in the prefix.

To avoid a crash for databases that use custom prefixes/suffixes like
those, instead of using the default journal, we create a temporary
journal to create the invoices in, which avoids these conflicts.

Task ID: 2251374

closes odoo/upgrade#1280

Signed-off-by: Christophe Simonis <chs@odoo.com>