[FIX] account@saas~12.4: bad account properties on res_partner
When creating `account.move` from draft/cancel `account.invoice`,
the `property_account_receivable_id` and `property_account_payable_id`
properties of res_partners may be used to create move lines.
If the accounts linked to these properties are not in the same company
than the properties themselves, the upgrade will fail.

To avoid that, bad properties of res_partners used in draft/cancel invoices are
removed, and a message is logged to the user.

Upgrade request: 47412

closes odoo/upgrade#1283

Signed-off-by: Denis Ledoux (dle) <dle@odoo.com>