[IMP] account@saas~12.3 : use CTE to avoid combinatorial explosion
Because we link tables to the updated one, not having it in the from
clause cause a combinatorial explosion. Using a with clause, and
properly joining records, generate a smaller result.

A side effect of the util.parallel_execute is a commit. As updating
twice a record in the same transaction is slower due to triggered
constraints since second update, this commit avoid this unwanted
behavior.

closes odoo/upgrade#1296

Signed-off-by: Christophe Simonis <chs@odoo.com>